[refactor] confirm page header component refactoring
